Please lint by using proper types
git-svn-id: file:///svn/unbound/trunk@4939 be551aaa-1e26-0410-a405-d3ace91eadb9

+++ util/config_file.h
@@ -163,7 +163,7 @@ struct config_file {
 	 * destinations */
 	int fast_server_permil;
 	/** number of fastest server to select from */
-	int fast_server_num;
+	size_t fast_server_num;
 
 	/** automatic interface for incoming messages. Uses ipv6 remapping,
 	 * and recvmsg/sendmsg ancillary data to detect interfaces, boolean */
